Update admin page display info component
