+++ packages/outside/src/components/AdminDisplayInfo/UserDisplayInfo/index.tsx
@@ -1,0 +1,14 @@
+import * as S from './style'
+
+const UserDisplayInfo = () => {
+  return (
+    <S.DisplayBar>
+      <span>이름</span>
+      <span>직업</span>
+      <span>전화번호</span>
+      <span>이메일</span>
+    </S.DisplayBar>
+  )
+}
+
+export default UserDisplayInfo
