+++ packages/outside/src/components/AdminDisplayInfo/index.ts
@@ -1,0 +1,3 @@
+export { default as UserDisplayInfo } from './UserDisplayInfo'
+export { default as NewDisplayInfo } from './NewDisplayInfo'
+export { default as WithdrawDisplayInfo } from './WithdrawDisplayInfo'
