+++ packages/outside/src/components/index.ts
@@ -1,3 +1,3 @@
 export { default as ClubItem } from './ClubItem'
 export { default as ClubBanner } from './ClubBanner'
-export { default as AdminDisplayInfo } from './AdminDisplayInfo'
+export * from './AdminDisplayInfo'
